// AppFormationFrontend/src/app/services/models/registration-request.ts
/* tslint:disable */
/* eslint-disable */
export interface RegistrationRequest {
  cin: string;
  company?: string;
  dateOfBirth?: string;
  email: string;
  firstname: string;
  lastname: string;
  password: string;
  pdfFile?: string;
  phoneNumber: string;
  speciality?: string;
  workplace?: string;
}